Add files via upload
Interaction data for over 9,000 accounts on SoundCloud, saved as a sparse matrix. Rows are labeled with remaining_seeds.csv, columns labeled with seeds.csv. remaining_seeds is a bit longer than the number of columns here, just cut off remaining_seeds beyond the number of rows in the matrix. Some rows may also be blank if few interactions were found or the link was broken.